fixing throttling assertions failures and addressed review comments
